Move to integer space for calculations.
This removes the obvious tear during the LB refrain ending, and fixes texture
seams I introduced in e42113b303b34ef62f4b463ec71c19aa55db33b1.

+++ src/Systems/SDL/Texture.hpp
@@ -100,7 +100,7 @@ class Texture {
     const Rect& src, const Rect& dst, const RGBAColour& rgba);
 
   bool filterCoords(int& x1, int& y1, int& x2, int& y2,
-                    float& dx1, float& dy1, float& dx2, float& dy2);
+                    int& dx1, int& dy1, int& dx2, int& dy2);
 
   int x_offset_;
   int y_offset_;
